#auth and config

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { TaskRepository } from './task.repository';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
 
 
 @Injectable()
@@ -15,13 +16,18 @@ export class TasksService {
   ) {}
   //private tasks: Task[] = [];
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]>{
-    return this.taskRepository.getTask(filterDto)
+  async getTasks(
+    filterDto: GetTasksFilterDto,
+    user: User,
+    ): Promise<Task[]>{
+    return this.taskRepository.getTask(filterDto, user)
   }
 
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(
+    id: number,
+    user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id }});
 
     if(!found) {
       throw new NotFoundException(`Task with ID "${id}" not found`);
@@ -32,19 +38,27 @@ export class TasksService {
 
 
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task>{
-    return this.taskRepository.createTask(createTaskDto)
+  async createTask(
+    createTaskDto: CreateTaskDto,
+    user: User,
+    ): Promise<Task>{
+    return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async updateStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateStatus(
+    id: number, 
+    status: TaskStatus,
+    user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
   }
 
-  async deleteTask(id: number): Promise<void>{
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(
+    id: number,
+    user: User): Promise<void>{
+    const result = await this.taskRepository.delete({ id, userId: user.id});
     if(result.affected === 0) {
       throw new NotFoundException(`Task with ID "${id}" not found`);
     }
